Bugfix #9 and #10
When selecting load game, the variable that stores whether the intro
scene had been watched wasn't being set. Because of this, the alternate
main menu wasn't being activated when pressing the Back buttons in the
sub-menus if the player had chosen Load Game instead of Play Game at the
primary main menu.